Clean up hybrid_engine features
