--- components/hybrid_engine/tests/failpoints/mod.rs
@@ -1,3 +1,0 @@
-// Copyright 2024 TiKV Project Authors. Licensed under Apache-2.0.
-
-mod test_write_batch;
